refactor(backend): extract ProviderRuntime from AgentService
Moves the provider domain out of agent-service.ts into src/backend/agent/provider-runtime.ts:
nineteen fields, the connect/login/refresh methods, #refreshModelCatalog, #probeComputerUse,
#requireReadyClient, #refreshUsage, #setStatus and #clientForBot, plus the declarations only
they used — INITIAL_STATUS, FALLBACK_MODELS, CURATED_CODEX_MODEL_IDS, the three login
timeouts and the two pending-login shapes. Everything moved is keyed by AgentProvider and
nothing else in the file is.

The ten public methods stay on AgentService as delegations, so Pick<AgentService, …> in
team-api-server.ts, the six IPC handler modules and the dynamic service[connect]() call in
providers.test.ts all need no edit.

Provider code called back into the core in three shapes, so the inversion is a ProviderHooks
record with four members rather than a shared base:

- bindClient wires #handleNotification and #handleServerRequest, which stay in the core
  because the notification router is where provider wire order becomes OpenBot state order.
- onProvidersReady is the tail #connect and #activateProviderClient each carried a copy of:
  reconcile unresolved deliveries, backfill provider history, schedule a drain per bot. Two
  copies of a four-line sequence that must stay in that order; now one.
- onProviderLost is the #handleExit cleanup — prompts, approvals, browser takeovers,
  compaction runtime, browser controls. The plan predicted this block would otherwise be
  rewritten by three later PRs and conflict three ways on the same twelve lines.
- isStopping reads the facade's flag, because stop() sets it before it disposes the runtime
  and a client exiting in that window must not trigger a restart.

#status was read outside the provider domain in five places and every one asked the same
question, so they get ProviderPort.isReady() rather than the status object. interruptAll,
#scheduleDrain, #drainBot and the compaction request path now read a boolean; #scheduleDrain
converting its own clause is the first instalment of the mute registry the plan describes,
and PRs 3, 5 and 6 convert theirs.

Three notification-router arms move behind named methods: completeCodexLogin,
setComputerUseCapability and refreshCodexUsage. The arms keep their position in the switch,
so the wire order the router encodes is unchanged.

Deviation from the plan, stated so review does not have to find it: there is no AgentContext.
As specified it holds `providers: ProviderPort` and `conversation: ConversationRuntime`, so a
ProviderRuntime that consumed it would depend on a record containing its own port. Controllers
take their dependencies explicitly instead, and ProviderPort is the narrow view later domains
receive. AgentContext is not deferred again — it is dropped, and each later PR names the ports
it needs.

Tests: all six agent-service shards untouched and green (91 tests), plus the two conversation
transaction tests. providers.test.ts covers this boundary already; PR 1's second commit
relocates its login and refresh cases to agent/provider-runtime.test.ts.

agent-service.ts: 5248 → 4160 lines.
